// src/app/models/datos.ts
export class Datosvictima {
  id: number;
  user_id: number;
  pais_code: string;
  numDatosvictimas: string;
  generoVictimasHombre: string;
  generoVictimasMujer: string;
  edadVictimaNino: string;
  edadVictimaJoven: string;
  edadVictimaAdulto: string;
  edadVictimaOld: string;
  estatusMigratorioRegular: string;
  estatusMigratorioIrregular: string;
  estatusConsulado: string;
  estadoIntegridad: string;
  created_at: Date;
}